requirements.txt
nonebot2[fastapi]
nonebot-adapter-onebot
jupyter
black[jupyter]
pydantic
python-dotenv
ujson
openai
wandb
weave
nb-cli
urlextract
